requirements.txt
telethon
pandas
pyyaml
pytesseract
Pillow